Apply other thresholds to DE results

Apply other thresholds to DE results

Usage

apply_thresholds(
  de_res,
  logFC = TRUE,
  logFC_up = 1,
  logFC_down = -1,
  p_adj = TRUE,
  alpha = 0.05
)

Arguments

de_res

data table resulting of run_DE

logFC

Boolean specifying whether to apply a logFC threshold (TRUE) or not (FALSE)

logFC_up

Upper log2 fold change threshold (dividing into up regulated)

logFC_down

Lower log2 fold change threshold (dividing into down regulated)

p_adj

Boolean specifying whether to apply a threshold on adjusted p-values (TRUE) or on raw p-values (FALSE)

alpha

Threshold for adjusted p-values or p-values

Value

data table updating the Change column with the newly applied thresholds

Examples

data(tuberculosis_TMT_de_res)
de_res <- apply_thresholds(tuberculosis_TMT_de_res, logFC = FALSE,
                           p_adj = TRUE, alpha = 0.01)