Outlier detection via POMA R Package

Outlier detection via POMA R Package

Usage

detect_outliers_POMA(
  se,
  ain = "log2",
  condition = NULL,
  method = "euclidean",
  type = "median",
  group = TRUE,
  coeff = 1.5
)

Arguments

se

SummarizedExperiment containing all necessary information of the proteomics data set

ain

String which data type should be used (default raw)

condition

Column name of condition (if NULL, condition saved in SummarizedExperiment will be taken)

method

String specifying the method that should be used to calculate the distance matrix

type

String specifying the type of distance calculation to centroid or spatial median

group

String specifying if the outlier detection should be performed multi-variate (with conditions) or on the complete data set

coeff

This value corresponds to the classical 1.5 in Q3 + 1.5 * IQR formula to detect outliers. By changing this value, the permissiveness in outlier detection will change.

Value

list of two ggplot objects and a data.table with outlier samples

Examples

data(tuberculosis_TMT_se)
poma_res <- detect_outliers_POMA(tuberculosis_TMT_se, ain="raw",
                                 condition = NULL, method="euclidean",
                                 type="median", group=TRUE, coeff = 1.5)
#> Condition of SummarizedExperiment used!
#> Warning: some squared distances are negative and changed to zero
#> Scale for fill is already present.
#> Adding another scale for fill, which will replace the existing scale.
#> Scale for colour is already present.
#> Adding another scale for colour, which will replace the existing scale.
#> Scale for fill is already present.
#> Adding another scale for fill, which will replace the existing scale.