Helper function to transform an expression data frame to a data table

Helper function to transform an expression data frame to a data table

Usage

express_to_DT(expr_data, column_names, row_names)

Arguments

expr_data

Expression data frame containing the expression data

column_names

Column names of the expression data

row_names

Row names of the expression data

Value

Data table containing the expression data