Extract consensus DE candidates

Extract consensus DE candidates

Usage

extract_consensus_DE_candidates(
  de_res,
  ain = NULL,
  comparisons = NULL,
  norm_thr = 0.8,
  per_comparison = FALSE
)

Arguments

de_res

data table resulting of run_DE

ain

Vector of strings of normalization methods to visualize (must be valid normalization methods saved in de_res)

comparisons

Vector of comparisons (must be valid comparisons saved in de_res)

norm_thr

Threshold for the number of normalization methods that must agree on a DE candidate

per_comparison

Logical indicating if the consensus should be calculated per comparison

Value

data table with consensus DE candidates

Examples

data(tuberculosis_TMT_de_res)
extract_consensus_DE_candidates(tuberculosis_TMT_de_res, ain = NULL,
                  comparisons = NULL, norm_thr = 0.8, per_comparison = TRUE)
#> # A tibble: 132 × 4
#> # Groups:   Protein.IDs [83]
#>    Protein.IDs                  Comparison Intersections Percentage
#>    <chr>                        <fct>              <int>      <dbl>
#>  1 A0A075B6I1                   TBL-HC                 5        100
#>  2 A0A075B6K4;P01715            TBL-HC                 5        100
#>  3 A0A075B6K4;P01715            Rx-TBL                 4         80
#>  4 A0A075B6K5;P80748;A0A075B6K2 TBL-HC                 5        100
#>  5 A0A075B6K5;P80748;A0A075B6K2 TBL-PTB                5        100
#>  6 A0A075B6K5;P80748;A0A075B6K2 Rx-TBL                 5        100
#>  7 A0A075B730;P58107            PTB-HC                 5        100
#>  8 A0A075B730;P58107            Rx-HC                  5        100
#>  9 A0A0A0MS08;P01857;A0A0A0MS07 Rx-HC                  5        100
#> 10 A0A0B4J1V6;A0A0B4J1V0        PTB-HC                 4         80
#> # ℹ 122 more rows