Normalize SummarizedExperiment object using combinations of normalization methods

Normalize SummarizedExperiment object using combinations of normalization methods

Usage

normalize_se_combination(
  se,
  methods,
  ains,
  on_raw = NULL,
  combination_pattern = "_on_",
  gamma.0 = 0.1,
  reduce_correlation_by = 1,
  NormicsVSN_quantile = 0.8,
  top_x = 50,
  VSN_quantile = 0.9
)

Arguments

se

SummarizedExperiment containing all necessary information of the proteomics data set

methods

Vector of normalization methods to apply for normalizing the proteomics data of the SummarizedExperiment object (identifier of normalization methods can be retrieved using get_all_normalization_methods())

ains

Vector of assays of SummarizedExperiment object to apply the normalization methods (e.g. if you want to perform Median normalization on IRS-normalized data)

on_raw

Logical indicating if the normalization should be performed on the raw data or on log2-transformed data. If on_raw = NULL(default), the normalization is performed on the default method specific on_raw setting (suggestion based on publications).

combination_pattern

String to give name to combination of methods (e.g. IRS_on_Median --> "_on_")

gamma.0

Numeric representing the exponent of the weighted density of RobNorm normalization. When the sample size is small, the fitted population of some proteins could be locally trapped such that the variance of those proteins was very small under a large gamma. To avoid this, a small gamma is recommended. When sample size smaller than 40, then set gamma to 0.5 or 0.1.

reduce_correlation_by

If the data is too big for the computation of the params, increase this parameter by 2,3,4.... The whole data will still be normalized, but the params are calculated on every second row etc.

NormicsVSN_quantile

The quantile that is used for the resistant least trimmed sum of squares regression. A value of 0.8 means focusing on the central 80% of the data, reducing the influence of outliers.

top_x

Number of reference proteins extracted for the calculation of parameters

VSN_quantile

Numeric of length 1. The quantile that is used for the resistant least trimmed sum of squares regression. (see vsn2 lts.quantile)

Value

SummarizedExperiment object with normalized data saved as assays

Examples

data(tuberculosis_TMT_se)
tuberculosis_TMT_se <- normalize_se_combination(tuberculosis_TMT_se,
          methods = c("Median","NormicsVSN"), ains = c("IRS"), on_raw = NULL,
          combination_pattern = "_on_", gamma.0 = 0.1,
          reduce_correlation_by = 1, NormicsVSN_quantile = 0.8, top_x = 50,
          VSN_quantile = 0.9)
#> IRS normalization not yet performed. Single IRS normalization performed now.
#> IRS completed.
#> Median normalization performed on IRS-normalized data completed.
#> NormicsVSN normalization performed on IRS-normalized data completed.