Run DE analysis of a selection of normalized data sets

Run DE analysis of a selection of normalized data sets

Usage

run_DE(
  se,
  comparisons,
  ain = NULL,
  condition = NULL,
  DE_method = "limma",
  covariate = NULL,
  logFC = TRUE,
  logFC_up = 1,
  logFC_down = -1,
  p_adj = TRUE,
  alpha = 0.05,
  B = 100,
  K = 500,
  trend = TRUE,
  robust = TRUE
)

Arguments

se

SummarizedExperiment containing all necessary information of the proteomics data set

comparisons

Vector of comparisons that are performed in the DE analysis (from specify_comparisons method)

ain

Vector of strings which assay should be used as input (default NULL). If NULL then all normalization of the se object are plotted next to each other.

condition

column name of condition (if NULL, condition saved in SummarizedExperiment will be taken)

DE_method

String specifying which DE method should be applied (limma, ROTS)

covariate

String specifying which column to include as covariate into limma

logFC

Boolean specifying whether to apply a logFC threshold (TRUE) or not (FALSE)

logFC_up

Upper log2 fold change threshold (dividing into up regulated)

logFC_down

Lower log2 fold change threshold (dividing into down regulated)

p_adj

Boolean specifying whether to apply a threshold on adjusted p-values (TRUE) or on raw p-values (FALSE)

alpha

Threshold for adjusted p-values or p-values

B

Number of bootstrapping for ROTS

K

Number of top-ranked features for reproducibility optimization

trend

logical, should an intensity-dependent trend be allowed for the prior variance? If FALSE then the prior variance is constant. Alternatively, trend can be a row-wise numeric vector, which will be used as the covariate for the prior variance.

robust

logical, should the estimation of df.prior and var.prior be robustified against outlier sample variances?

Value

Data table of DE results of selected normalized data sets

Examples

data(tuberculosis_TMT_se)
comparisons <- specify_comparisons(tuberculosis_TMT_se, condition = NULL,
                            sep = NULL, control = NULL)
#> Condition of SummarizedExperiment used!
de_res <- run_DE(tuberculosis_TMT_se, comparisons,
              ain = NULL, condition = NULL, DE_method = "limma",
              logFC = TRUE, logFC_up = 1, logFC_down = -1, p_adj = TRUE,
              alpha = 0.05, B = 100, K = 500, trend = TRUE, robust = TRUE)
#> Condition of SummarizedExperiment used!
#> All assays of the SummarizedExperiment will be used.
#> DE Analysis will not be performed on raw data.
#> Warning: Partial NA coefficients for 114 probe(s)
#> log2: DE analysis completed.
#> Warning: Partial NA coefficients for 114 probe(s)
#> RobNorm: DE analysis completed.
#> Warning: Partial NA coefficients for 112 probe(s)
#> IRS_on_RobNorm: DE analysis completed.
#> Warning: Partial NA coefficients for 114 probe(s)
#> Median: DE analysis completed.
#> Warning: Partial NA coefficients for 112 probe(s)
#> IRS_on_Median: DE analysis completed.